+++ ark_nlp/dataset/__init__.py
@@ -4,8 +4,8 @@ from ark_nlp.dataset.base._sentence_classification_dataset import SentenceClassi
 from ark_nlp.dataset.base._sentence_classification_dataset import SentenceClassificationDataset as SCDataset
 from ark_nlp.dataset.base._sentence_classification_dataset import PairMergeSentenceClassificationDataset
 from ark_nlp.dataset.base._sentence_classification_dataset import PairMergeSentenceClassificationDataset as PMSCDataset
-from ark_nlp.dataset.base._sentence_classification_dataset import TwinTowersSentenceClassificationDataset
-from ark_nlp.dataset.base._sentence_classification_dataset import TwinTowersSentenceClassificationDataset as TTSCDataset
+from ark_nlp.dataset.base._sentence_classification_dataset import PairWiseSentenceClassificationDataset
+from ark_nlp.dataset.base._sentence_classification_dataset import PairWiseSentenceClassificationDataset as PWSCDataset
 
 from ark_nlp.dataset.base._token_classification_dataset import TokenClassificationDataset
