--- ark_nlp/model/__init__.py
@@ -1,1 +1,0 @@
-# from ark_nlp.model.bert_sequence_classification import BertSequenceClassificationModel
