+++ ark_nlp/nn/__init__.py
@@ -9,11 +9,6 @@ from ark_nlp.nn.base.ernie import Ernie
 from ark_nlp.nn.base.nezha import NeZha
 from ark_nlp.nn.base.roformer import RoFormer
 
-from ark_nlp.nn.biaffine_bert import BiaffineBert
-from ark_nlp.nn.span_bert import SpanBert
-from ark_nlp.nn.global_pointer_bert import GlobalPointerBert
-from ark_nlp.nn.crf_bert import CrfBert
-
 from ark_nlp.nn.prompt_bert import BertForPromptMaskedLM
 
 from transformers import BertConfig
