--- ark_nlp/nn/crf_bert.py
@@ -1,41 +1,0 @@
-# Copyright (c) 2021 DataArk Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Author: Xiang Wang, xiangking1995@163.com
-# Status: Active
-
-
-from ark_nlp.nn.base.bert import BertForTokenClassification
-from ark_nlp.nn.layer.crf_block import CRF
-
-
-class CrfBert(BertForTokenClassification):
-    """
-    基于BERT + CRF 的命名实体模型
-
-    Args:
-        config: 
-            模型的配置对象
-        encoder_trained (:obj:`bool`, optional, defaults to True):
-            bert参数是否可训练，默认可训练
-    """  # noqa: ignore flake8"
-
-    def __init__(
-        self,
-        config,
-        encoder_trained=True
-    ):
-        super(CrfBert, self).__init__(config, encoder_trained)
-
-        self.crf = CRF(num_tags=config.num_labels, batch_first=True)
