--- ark_nlp/nn/global_pointer_bert.py
@@ -1,83 +1,0 @@
-# Copyright (c) 2021 DataArk Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Author: Xiang Wang, xiangking1995@163.com
-# Status: Active
-
-
-from transformers import BertModel
-from transformers import BertPreTrainedModel
-from ark_nlp.nn.layer.global_pointer_block import GlobalPointer
-
-
-class GlobalPointerBert(BertPreTrainedModel):
-    """
-    基于GlobalPointe的命名实体模型
-
-    Args:
-        config: 
-            模型的配置对象
-        encoder_trained (:obj:`bool`, optional, defaults to True):
-            bert参数是否可训练，默认可训练
-        head_size (:obj:`int`, optional, defaults to 64):
-            GlobalPointer head个数
-
-    Reference:
-        [1] https://www.kexue.fm/archives/8373
-        [2] https://github.com/suolyer/PyTorch_BERT_Biaffine_NER
-    """  # noqa: ignore flake8"
-
-    def __init__(
-        self,
-        config,
-        encoder_trained=True,
-        head_size=64
-    ):
-        super(GlobalPointerBert, self).__init__(config)
-
-        self.num_labels = config.num_labels
-
-        self.bert = BertModel(config)
-
-        for param in self.bert.parameters():
-            param.requires_grad = encoder_trained
-
-        self.global_pointer = GlobalPointer(
-            self.num_labels,
-            head_size,
-            config.hidden_size
-        )
-
-        self.init_weights()
-
-    def forward(
-        self,
-        input_ids=None,
-        attention_mask=None,
-        token_type_ids=None,
-        **kwargs
-    ):
-        outputs = self.bert(
-            input_ids,
-            attention_mask=attention_mask,
-            token_type_ids=token_type_ids,
-            return_dict=True,
-            output_hidden_states=True
-        ).hidden_states
-
-        sequence_output = outputs[-1]
-
-        logits = self.global_pointer(sequence_output, mask=attention_mask)
-
-        return logits
